+++ src/storage/index.ts
@@ -1,7 +1,9 @@
 import { disallowPublicAccess } from "./disallowPublicAccess";
 import { requireUniformBucketLevelAccess } from "./requireUniformBucketLevelAccess";
+import { requireBucketVersioning } from "./requireBucketVersioning";
 
 export const storagePolicies = [
 	disallowPublicAccess,
+	requireBucketVersioning,
 	requireUniformBucketLevelAccess,
 ];