+++ src/storage/index.ts
@@ -2,6 +2,7 @@ import { disallowPublicAccess } from "./disallowPublicAccess";
 import { disallowSelfBucketLogging } from "./disallowSelfBucketLogging";
 import { requireBucketLogging} from "./requireBucketLogging";
 import { requireBucketVersioning } from "./requireBucketVersioning";
+import { requireCmek } from "./requireCmek";
 import { requireUniformBucketLevelAccess } from "./requireUniformBucketLevelAccess";
 
 export const storagePolicies = [
@@ -9,5 +10,6 @@ export const storagePolicies = [
 	disallowSelfBucketLogging,
 	requireBucketLogging,
 	requireBucketVersioning,
+	requireCmek,
 	requireUniformBucketLevelAccess,
 ];