+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cloud Storage policies (#9):
   - `disallowPublicBuckets`
   - `requireUniformBucketLevelAccess`
+  - `requireBucketVersioning`
 
 ## [1.2.0] - 2023-12-27