+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `disallowSelfBucketLogging`
   - `requireUniformBucketLevelAccess`
   - `requireBucketVersioning`
+  - `requireBucketLogging`
+  - `requireCmek`
 
 ## [1.2.0] - 2023-12-27