+++ CHANGELOG.md
@@ -12,10 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cloud Storage policies (#9):
   - `disallowPublicBuckets`
   - `disallowSelfBucketLogging`
-  - `requireUniformBucketLevelAccess`
-  - `requireBucketVersioning`
   - `requireBucketLogging`
+  - `requireBucketVersioning`
   - `requireCmek`
+  - `requireUniformBucketLevelAccess`
 
 ## [1.2.0] - 2023-12-27