+++ CHANGELOG.md
@@ -17,8 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `requireCmek`
   - `requireUniformBucketLevelAccess`
 - BigQuery policies (#10):
-  - `requireCmekDataset`
-  - `requireCmekTable`
+  - `datasetDeletionProtection`
+  - `datasetRequireCmek`
+  - `tableDeletionProtection`
+  - `tableRequireCmek`
 
 ## [1.2.0] - 2023-12-27