+++ CHANGELOG.md
@@ -18,8 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `requireUniformBucketLevelAccess`
 - BigQuery policies (#10):
   - `datasetDeletionProtection`
+  - `datasetDisallowPublicAccess`
   - `datasetRequireCmek`
   - `tableDeletionProtection`
+  - `tableDisallowPublicAccess`
   - `tableRequireCmek`
 
 ## [1.2.0] - 2023-12-27